The root folder is now kept track of as a string, which is less confusing
We are about to rely on a specified path AT RUNTIME to figure out where to load Resources from. If we keep track of that separately, then you will get the confusing behaviour that if the specified folder is moved, things will work fine in the editor, but they will break in builds. I would rather then also have it break in the editor so that you're aware that it's a problem and will fix it.